start cleaning the code
